refactor: move mmio_read_bytes to klibc and use PciCpuAddr for fb_base
- Move mmio_read_bulk from devfs into klibc::mmio::read_bytes (generic utility)
- Change fb_base() to return Option<PciCpuAddr> instead of raw usize
- Use proper PciCpuAddr address type throughout DevFramebuffer read/write
